% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_type.R
\name{fixer_account_type}
\alias{fixer_account_type}
\title{Set account type}
\usage{
fixer_account_type(type = c("free", "paid"))
}
\arguments{
\item{type}{Your \href{https://fixer.io/}{fixer.io} account type. Accepts
one of \code{"free"} or \code{"paid"}.}
}
\description{
Paid accounts have access to additional features.
\code{fixerapi} uses this information to activate SSL encryption.

The \href{https://fixer.io/}{fixer.io} API still checks the
account type when receiving requests, so setting you account type here only
forces HTTPS instead of HTTP. If you have a paid account but do not set
your account status you will not be able to use the HTTPS connection but
there will be no other affect on your usage of the API.
}
