% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{fixer_api_key}
\alias{fixer_api_key}
\title{Fixer API Key}
\usage{
fixer_api_key(force = FALSE)
}
\arguments{
\item{force}{If TRUE, resets the API key and requires a
new key to be provided.}
}
\description{
Fixer API Key
}
