% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{fixer_convert}
\alias{fixer_convert}
\title{Convert currencies}
\usage{
fixer_convert(from, to, amount = 1, date = NULL)
}
\arguments{
\item{from}{Symbol of currency to convert from.}

\item{to}{Symbol of currency to convert to.}

\item{amount}{Value of currency to convert. Defaults to 1.}

\item{date}{Optional variable. A date in YYYY-MM-DD format, or any value
that can be coerced to YYYY-MM-DD format with \code{as.Date()}.
Defaults to \code{NULL}, which returns the latest conversion data.}
}
\value{
A tibble with the value of the conversion
}
\description{
Currency conversion is only available on paid plans.
}
\examples{

}
