% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{fixer_latest}
\alias{fixer_latest}
\title{Latest currency data available}
\usage{
fixer_latest(base = "EUR", symbols = NULL)
}
\arguments{
\item{base}{The base currency to index other currencies against.
Defaults to \code{"EUR"}. Can only be changed with paid plans.}

\item{symbols}{The symbols of currencies to return. Defaults to \code{NULL}
and returns all available currencies. See \code{\link{fixer_symbols}} for
details on symbol options.}
}
\description{
Returns a tibble with the most recently available currency conversion data
available from the Fixer API
}
