% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{fixer_timeseries}
\alias{fixer_timeseries}
\title{Exchange rate time series}
\usage{
fixer_timeseries(start_date, end_date, base = "EUR", symbols = NULL)
}
\arguments{
\item{start_date}{The start date of requested timeseries.}

\item{end_date}{The end date of requested timeseries.}

\item{base}{The base currency to index other currencies against.
Defaults to \code{"EUR"}. Can only be changed with paid plans.}

\item{symbols}{The symbols of currencies to return exchange rates for.
Defaults to \code{NULL} and returns all available currencies. See
\code{\link{fixer_symbols}} for details on symbol options.}
}
\description{
Exchange rate time series
}
