% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixerapi-package.R
\docType{package}
\name{fixerapi}
\alias{fixerapi}
\alias{fixerapi-package}
\title{fixerapi: an R client for the fixer.io currency api}
\description{
The \href{https://fixer.io}{fixer.io} API provides real-time
exchange rates for 170 currencies, covering 1999 to the present.
\href{https://fixer.io}{fixer.io} requires an account and a private key,
and most features require a \href{https://fixer.io/product}{paid account}.
See the \href{https://fixer.io/documentation}{full API documentation} for
more details on account features and pricing.
}
