% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_status.R
\name{fixer_account_type}
\alias{fixer_account_type}
\title{Set account type}
\usage{
fixer_account_type(type = c("free", "basic", "pro", "plus"))
}
\arguments{
\item{type}{fixer.io account type. Accepts one of \code{"free"},
\code{"basic"}, \code{"pro"} or \code{"plus"}.}
}
\description{
Paid accounts have access to additional features. \code{fixerrr} uses this
information to activate paid features like SSL encryption and prevent API
calls that are not allowed by a users account type.
}
